+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 # 1.6.3 
 - UV Transform texture preview respect closest filtering
 - Fix switching to the Level Design workspace incorrectly applying textures in some cases
-- Default edge length on
+- Default show edge length on
+- Add show edge length, subdivisions, and Unit system to addon preferences
 
 # 1.6.2
 - Improve back face selection culling edge / vert selection (fix bugs and more heuristics in ambiguous situations)